requirements-bot.txt
python-dotenv==0.5.0
python-telegram-bot==4.1.2
PyYAML==3.11
.
